+++ compose/src/main/kotlin/com/yourssu/handy/compose/textfield/TextField.kt
@@ -74,45 +74,3 @@ fun TextField(
         }
     }
 }
-
-
-@Preview
-@Composable
-fun TextFieldPreview() {
-    HandyTheme {
-        Column(
-            modifier = Modifier.wrapContentSize().padding(20.dp),
-            verticalArrangement = Arrangement.spacedBy(20.dp)
-        ) {
-            val (text, onValueChange) = remember { androidx.compose.runtime.mutableStateOf("") }
-            TextField(
-                value = text,
-                textLabel = "Label",
-                helperText = "Helper Text",
-                placeholder = "placeholder",
-                trailingIcon = HandyIcons.Filled.Cancel,
-                onValueChange = onValueChange
-            )
-
-            TextField(
-                value = text,
-                textLabel = "Label",
-                helperText = "Helper Text",
-                placeholder = "placeholder",
-                trailingIcon = HandyIcons.Filled.Cancel,
-                isError = true,
-                onValueChange = onValueChange
-            )
-
-            TextField(
-                value = text,
-                textLabel = "Label",
-                helperText = "Helper Text",
-                placeholder = "placeholder",
-                trailingIcon = HandyIcons.Filled.Cancel,
-                enabled = false,
-                onValueChange = onValueChange
-            )
-        }
-    }
-}
